Move FavoriteFirestoreDatabase::getCurrentUser as RxFirebaseAuth
